Create starting dirs and README files
This is a beginning point on where to start for each piece of the
application. The dir structure will likely change, but I need to begin
somewhere. Also, I intend to document everything as I go along in README.md
files as necessary. A separate "doc" diretory might be created later. the
Javascript and Python README.md files start with some ideas on getting the
code going.