Functional Hashcash
Hashcash verification is now functional, although very, very slow. On my
observations, it takes about a full minute to search through a full
20-bits. On average, it should be taking maybe a second or two max. I think
this is because I'm operating on strings, and should operate on binary
objects instead. I need to speed up the while loop, by at least an order of
10.