Reworking of code structure and web.py starts
Mostly a complete revamp of how the layout should be for holding the code.
The idea is that the user can download a tarball, untar it to the web root
directory, make the necessary web server changes, and launch the site.
Also, this marks the beginning of developing against the web.py framework.
None of the code is working at the moment. Just a skeleton to get going.