Rename of some variables. Always use salts.
Rather than generate a 112-byte string from PBKDF2, the AES and HMAC keys are
generated from the nonce that creates the URL and the AES and HMAC salts
provided in dconfig.py. As a result of this change, some variables have been
renamed for better self-documentation.

Also, to help with additional code self-documentation the 'key' variable has
been renamed to 'aes_key'.