Add back the IV as the initial value for CTR.
Initially, I had added this with only 8-bytes (see commit 59f5b6e), rather than
the 12-bytes currently being committed. I was worried that I would run out of
counters before finishing the encryption of the data. Turns out, I have plenty
of room to spare.

If the counting space is setup to 128-bits, and the IV space is setup to
64-bits, and suppose the initial value of the counter is 2^64, then this leaves
64-bits of counters, counting 16-bytes at a time, or about 2^68 total bytes that
can be encrypted. 2^32 bytes is 4 GB.

As such, the new IV space is 96-bits, or 12 bytes. If the IV is actually 2^96,
then this would leave us with:

    2^32 counters * 16 bytes == 2^36 total bytes that can be encrypted.

I don't ever plan on encrypting more than 8 MB, which is 2^23 bytes. We're good.