Fixes #39.
The "setup.py" script will not create a "/data/" directory in the root path of
the web application, if it doesn't already exist, and will also create a
"dconfig.py" with the necessary keys and values. As such, the logic for this was
taken out of "note.py" and the "INSTALL.md" file has been update to reflect
getting the application setup correctly.